Create new component for display preferences
