+++ src/components/UserMenu.jsx
@@ -1,17 +1,16 @@
 import React from 'react';
 import createPersistedState from 'use-persisted-state';
 
-import { Add, ArrowDropDown, DarkMode, Done, Edit, Fullscreen, Person, RemoveCircleOutline, SettingsBrightness, WbSunny } from '@mui/icons-material';
-import { Box, Button, Chip, Divider, IconButton, List, ListItem, ListItemIcon, ListItemText, Menu, Switch, TextField, ToggleButton, ToggleButtonGroup, Typography } from '@mui/material';
+import { Add, ArrowDropDown, Done, Edit, Person, RemoveCircleOutline } from '@mui/icons-material';
+import { Box, Button, Chip, Divider, IconButton, List, ListItem, ListItemText, Menu, TextField, Typography } from '@mui/material';
 import makeStyles from '@mui/styles/makeStyles';
 
 import { useParams } from 'react-router-dom/cjs/react-router-dom.min';
 import useSpots from '../hooks/useSpots';
 import SpotDialog from './SpotDialog';
+import PreferencesDisplay from './PreferencesDisplay';
 
 const useTriState = createPersistedState('tri');
-const useThemePrefs = createPersistedState('themePref');
-const useMaxWidthState = createPersistedState('useMaxWidth');
 const useFavoritesState = createPersistedState('favorites');
 
 const useStyles = makeStyles(theme => {
@@ -21,13 +20,6 @@ const useStyles = makeStyles(theme => {
   return {
     icon: {
       [minWidth('md')]: { display: 'none' },
-    },
-    themeIcon: {
-      marginRight: 7,
-      fill: 'currentColor',
-    },
-    themeLabel: {
-      textTransform: 'none',
     },
     text: {
       textTransform: 'none',
@@ -39,9 +31,7 @@ const useStyles = makeStyles(theme => {
 
 const UserMenu = () => {
   const [tri, setTri] = useTriState();
-  const [useMaxWidth, setUseMaxWidth] = useMaxWidthState();
   const [favorites, setFavorites] = useFavoritesState([]);
-  const [themePrefs, setThemePrefs] = useThemePrefs('system');
   const [disableInput, setDisableInput] = React.useState(true);
   const [textValue, setTextValue] = React.useState(tri);
 
@@ -117,40 +107,8 @@ const UserMenu = () => {
         <Typography style={{ paddingLeft: 10, paddingTop: 10 }} gutterBottom variant="h6">
           Préferences
         </Typography>
-        <Divider textAlign="left">Affichage</Divider>
-        <List dense>
-          <ListItem>
-            <ToggleButtonGroup
-              size="small"
-              value={themePrefs}
-              exclusive
-              fullWidth
-            >
-              <ToggleButton onClick={() => setThemePrefs('dark')} value="dark">
-                <DarkMode className={classes.themeIcon} />
-                <span className={classes.themeLabel}>Sombre</span>
-              </ToggleButton>
-              <ToggleButton onClick={() => setThemePrefs('system')} value="system">
-                <SettingsBrightness className={classes.themeIcon} />
-                <span className={classes.themeLabel}>Système</span>
-              </ToggleButton>
-              <ToggleButton onClick={() => setThemePrefs('light')} value="light">
-                <WbSunny className={classes.themeIcon} />
-                <span className={classes.themeLabel}>Clair</span>
-              </ToggleButton>
-            </ToggleButtonGroup>
-          </ListItem>
-          <ListItem>
-            <ListItemIcon sx={{ marginRight: '-25px' }}><Fullscreen /></ListItemIcon>
-            <ListItemText primary="Pleine largeur" />
-            <Switch
-              checked={useMaxWidth}
-              onChange={() => {
-                setUseMaxWidth(!useMaxWidth);
-              }}
-            />
-          </ListItem>
-        </List>
+
+        <PreferencesDisplay />
 
         <Divider textAlign="left">
           Postes Favoris
@@ -165,7 +123,6 @@ const UserMenu = () => {
             onClick={() => setDialogOpen(!dialogOpen)}
           />
         </Divider>
-
         <List dense sx={{ pb: '12px' }}>
           {favorites.length === 0 && (
             <Typography sx={{ opacity: 0.4, textAlign: 'center', fontSize: '12px', margin: '15px' }}>
@@ -203,6 +160,7 @@ const UserMenu = () => {
             );
           })}
         </List>
+
         <Divider textAlign="left">Trigramme</Divider>
         <Box sx={{ display: 'flex', alignItems: 'flex-end', mx: '16px', my: '10px' }}>
           <TextField
